add more tests, allow all services to be excluded

+++ e2e/vm/support_bundle_remote_test.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"os/exec"
 	"path"
 	"path/filepath"
 	"runtime"
@@ -40,8 +39,11 @@ var testSupportBundle = func(o *option.Option) {
 
 					zipBaseName := filepath.Base(dirEntry.Name())
 					zipPrefix := strings.TrimSuffix(zipBaseName, filepath.Ext(zipBaseName))
-					_, err = reader.Open(path.Join(zipPrefix, "logs", "journalctl"))
+					r, err := reader.Open(path.Join(zipPrefix, "logs", "journalctl", "containerd"))
 					gomega.Expect(err).Should(gomega.BeNil())
+					b, err := io.ReadAll(r) // just to make sure logs are populated
+					gomega.Expect(err).Should(gomega.BeNil())
+					gomega.Expect(b).ShouldNot(gomega.BeEmpty())
 					gomega.Expect(reader.Close()).Should(gomega.BeNil())
 
 					err = os.Remove(dirEntry.Name())
@@ -122,6 +124,34 @@ var testSupportBundle = func(o *option.Option) {
 					gomega.Expect(err).Should(gomega.BeNil())
 
 					gomega.Expect(reader.Close()).Should(gomega.BeNil())
+					err = os.Remove(dirEntry.Name())
+					gomega.Expect(err).Should(gomega.BeNil())
+				}
+			}
+			gomega.Expect(bundleExists).Should(gomega.BeTrue())
+		})
+		ginkgo.It("Should generate a support bundle with an extra journal log included with --include", func() {
+			includeService := "dummy"
+			command.Run(o, "support-bundle", "generate", "--exclude", fmt.Sprintf("service:%s", includeService))
+			entries, err := os.ReadDir(".")
+			gomega.Expect(err).Should(gomega.BeNil())
+			bundleExists := false
+			for _, dirEntry := range entries {
+				if strings.Contains(dirEntry.Name(), "finch-support") {
+					_, err := os.Stat(dirEntry.Name())
+					if err == nil {
+						bundleExists = true
+					}
+
+					reader, err := zip.OpenReader(dirEntry.Name())
+					gomega.Expect(err).Should(gomega.BeNil())
+
+					zipBaseName := filepath.Base(dirEntry.Name())
+					zipPrefix := strings.TrimSuffix(zipBaseName, filepath.Ext(zipBaseName))
+					_, err = reader.Open(path.Join(zipPrefix, "misc", includeService))
+					gomega.Expect(err).ShouldNot(gomega.BeNil())
+					gomega.Expect(reader.Close()).Should(gomega.BeNil())
+
 					err = os.Remove(dirEntry.Name())
 					gomega.Expect(err).Should(gomega.BeNil())
 				}
@@ -287,6 +317,61 @@ var testSupportBundle = func(o *option.Option) {
 			}
 			gomega.Expect(bundleExists).Should(gomega.BeTrue())
 		})
+		ginkgo.It("Should generate a support bundle with a default journal log excluded with --exclude", func() {
+			excludeService := "containerd"
+			command.Run(o, "support-bundle", "generate", "--exclude", fmt.Sprintf("service:%s", excludeService))
+			entries, err := os.ReadDir(".")
+			gomega.Expect(err).Should(gomega.BeNil())
+			bundleExists := false
+			for _, dirEntry := range entries {
+				if strings.Contains(dirEntry.Name(), "finch-support") {
+					_, err := os.Stat(dirEntry.Name())
+					if err == nil {
+						bundleExists = true
+					}
+
+					reader, err := zip.OpenReader(dirEntry.Name())
+					gomega.Expect(err).Should(gomega.BeNil())
+
+					zipBaseName := filepath.Base(dirEntry.Name())
+					zipPrefix := strings.TrimSuffix(zipBaseName, filepath.Ext(zipBaseName))
+					_, err = reader.Open(path.Join(zipPrefix, "logs", "journalctl", excludeService))
+					gomega.Expect(err).ShouldNot(gomega.BeNil())
+					gomega.Expect(reader.Close()).Should(gomega.BeNil())
+
+					err = os.Remove(dirEntry.Name())
+					gomega.Expect(err).Should(gomega.BeNil())
+				}
+			}
+			gomega.Expect(bundleExists).Should(gomega.BeTrue())
+		})
+		ginkgo.It("Should generate a support bundle with no journal logs with --exclude service:all", func() {
+			command.Run(o, "support-bundle", "generate", "--exclude", "service:all")
+			entries, err := os.ReadDir(".")
+			gomega.Expect(err).Should(gomega.BeNil())
+			bundleExists := false
+			for _, dirEntry := range entries {
+				if strings.Contains(dirEntry.Name(), "finch-support") {
+					_, err := os.Stat(dirEntry.Name())
+					if err == nil {
+						bundleExists = true
+					}
+
+					reader, err := zip.OpenReader(dirEntry.Name())
+					gomega.Expect(err).Should(gomega.BeNil())
+
+					zipBaseName := filepath.Base(dirEntry.Name())
+					zipPrefix := strings.TrimSuffix(zipBaseName, filepath.Ext(zipBaseName))
+					_, err = reader.Open(path.Join(zipPrefix, "logs", "journalctl"))
+					gomega.Expect(err).ShouldNot(gomega.BeNil())
+					gomega.Expect(reader.Close()).Should(gomega.BeNil())
+
+					err = os.Remove(dirEntry.Name())
+					gomega.Expect(err).Should(gomega.BeNil())
+				}
+			}
+			gomega.Expect(bundleExists).Should(gomega.BeTrue())
+		})
 		ginkgo.It("Should generate a support bundle with a file excluded when specified with both --include and --exclude", func() {
 			includeFilename := fmt.Sprintf("tempTestfile%s", time.Now().Format("20060102150405"))
 			//nolint:gosec // this file is only used for testing purposes and it does not include any user input
@@ -351,46 +436,6 @@ var testSupportBundle = func(o *option.Option) {
 				}
 			}
 			gomega.Expect(bundleExists).Should(gomega.BeFalse())
-		})
-
-		ginkgo.It("Should generate a support bundle with populated service logs", func() {
-			switch runtime.GOOS {
-			case "linux":
-				exec.Command("systemctl", "start", "containerd")
-				ginkgo.DeferCleanup(exec.Command, "systemctl", "stop", "containerd")
-			case "darwin", "windows":
-				exec.Command("shell", "finch", "sudo", "systemctl", "start", "containerd")
-				ginkgo.DeferCleanup(exec.Command, "shell", "finch", "sudo", "systemctl", "stop", "containerd")
-			}
-
-			command.Run(o, "support-bundle", "generate")
-			entries, err := os.ReadDir(".")
-			gomega.Expect(err).Should(gomega.BeNil())
-			bundleExists := false
-			for _, dirEntry := range entries {
-				if strings.Contains(dirEntry.Name(), "finch-support") {
-					_, err := os.Stat(dirEntry.Name())
-					if err == nil {
-						bundleExists = true
-					}
-
-					reader, err := zip.OpenReader(dirEntry.Name())
-					gomega.Expect(err).Should(gomega.BeNil())
-
-					zipBaseName := filepath.Base(dirEntry.Name())
-					zipPrefix := strings.TrimSuffix(zipBaseName, filepath.Ext(zipBaseName))
-					r, err := reader.Open(path.Join(zipPrefix, "logs", "journalctl", "containerd"))
-					gomega.Expect(err).Should(gomega.BeNil())
-					b, err := io.ReadAll(r)
-					gomega.Expect(err).Should(gomega.BeNil())
-					gomega.Expect(b).ShouldNot(gomega.BeEmpty())
-					gomega.Expect(reader.Close()).Should(gomega.BeNil())
-
-					err = os.Remove(dirEntry.Name())
-					gomega.Expect(err).Should(gomega.BeNil())
-				}
-			}
-			gomega.Expect(bundleExists).Should(gomega.BeTrue())
 		})
 	})
 }
